Single responsibility object that increments session step
